requirements.txt
aiogram>=3.4,<4.0
aiosqlite>=0.19
qrcode[pil]>=7.4
Pillow>=10.0
openpyxl>=3.1
